requirements.txt
baidu-aip
idna
pyaudio==0.2.11
speechrecognition
urllib3
transformers==2.1.1
torch==1.3.1
falcon==1.2.0
inflect==0.2.5
audioread==2.1.5
librosa==0.5.1
matplotlib==2.0.2
numpy==1.14.0
scipy==1.0.0
tqdm==4.11.2
Unidecode==0.4.20
pyaudio==0.2.11
sounddevice==0.3.10
lws==1.2.0
keras